MKStoreManager.m: fix signed/unsigned comparison warning

+++ MKStoreManager.m
@@ -252,7 +252,7 @@ static MKStoreManager* _sharedStoreManager;
 	[self.purchasableObjects addObjectsFromArray:response.products];
 	
 #ifndef NDEBUG	
-	for(int i=0;i<[self.purchasableObjects count];i++)
+	for(NSUInteger i=0;i<[self.purchasableObjects count];i++)
 	{		
 		SKProduct *product = [self.purchasableObjects objectAtIndex:i];
 		NSLog(@"Feature: %@, Cost: %f, ID: %@",[product localizedTitle],
@@ -293,7 +293,7 @@ static MKStoreManager* _sharedStoreManager;
 - (NSMutableArray*) purchasableObjectsDescription
 {
 	NSMutableArray *productDescriptions = [[NSMutableArray alloc] initWithCapacity:[self.purchasableObjects count]];
-	for(int i=0;i<[self.purchasableObjects count];i++)
+	for(NSUInteger i=0;i<[self.purchasableObjects count];i++)
 	{
 		SKProduct *product = [self.purchasableObjects objectAtIndex:i];
 		
@@ -326,7 +326,7 @@ NSString *upgradePrice = [prices objectForKey:@"com.mycompany.upgrade"]
 */
 - (NSMutableDictionary *)pricesDictionary {
     NSMutableDictionary *priceDict = [NSMutableDictionary dictionary];
-	for(int i=0;i<[self.purchasableObjects count];i++)
+	for(NSUInteger i=0;i<[self.purchasableObjects count];i++)
 	{
 		SKProduct *product = [self.purchasableObjects objectAtIndex:i];
